refactor(frontend): extract Gantt view-state helpers into pages/ganttChartState
Move the pure helpers from pages/GanttChart.tsx — persisted view/state
storage, timeline date and scroll math, row sizing constants, and the
synthetic touch-to-mouse event utilities — verbatim into
pages/ganttChartState.ts. The page component and the
GanttRenderBoundary error boundary stay in GanttChart.tsx. No behavior
changes.
